Creates a custom parameters class
